Add support for link type between issues (#964)
* Add support for link type between issues

Fixes #847

* Create AbstractIssue
