+++ src/test/resources/org/gitlab4j/api/linked-issues.json
@@ -22,6 +22,9 @@
     "subscribed" : true,
     "user_notes_count": 0,
     "web_url": "http://example.com/example/example/issues/14",
-    "confidential": false
+    "confidential": false,
+    "link_type": "relates_to",
+    "link_created_at": "2016-01-07T12:44:33.959Z",
+    "link_updated_at": "2016-01-07T12:44:33.959Z"
   }
 ]
